example of divided resourses
I think that up to 8 VPSs dividing up the resources each member has is a potentially interesting feature for some niche users.  e.g. if you are using the VPS to host websites, you can have one VPS for normal, run of-the-mill websites, and then split off resources and run a dedicated VPS for someones political campaign, which might be more likely a target of hackers, so if they manage to take down the political campaign server the rest of your hosted sites are unaffected...